@@ -1,13 +1,15 @@
 package reforged.marcin.krysiak.basketstats.models;
 
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.Set;
 
 @Getter
 @Setter
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
 @SequenceGenerator(name="seq", initialValue=10)
 @Table(name = "leagues")
@@ -21,14 +23,5 @@ public class League {
 
     @Column(length = 50, nullable = false)
     private String name;
-
-    public League() {
-
-    }
-
-    public League(Long id, String name, Set<Team> teams) {
-        this.id = id;
-        this.name = name;
-    }
 
 }
