@@ -1,9 +1,8 @@
 package reforged.marcin.krysiak.basketstats.models;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-
-import javax.persistence.*;
 
 @Getter
 @Setter
