@@ -1,9 +1,9 @@
 package reforged.marcin.krysiak.basketstats.models;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
 import java.util.Arrays;
 
 @Getter
