@@ -1,6 +1,9 @@
 package reforged.marcin.krysiak.basketstats.service.impl;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import reforged.marcin.krysiak.basketstats.models.League;
@@ -11,11 +14,15 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-@Transactional
+@RequiredArgsConstructor
 public class LeagueServiceImpl implements LeagueService {
 
-    @Autowired
-    LeagueRepository leagueRepository;
+    private final LeagueRepository leagueRepository;
+
+    @Override
+    public Page<League> getLeaguesBySpec(Specification<League> spec, Pageable pageable) {
+        return leagueRepository.findAll(spec, pageable);
+    }
 
     @Override
     public List<League> getAllLeagues() {
@@ -33,12 +40,14 @@ public class LeagueServiceImpl implements LeagueService {
     }
 
     @Override
+    @Transactional
     public League createLeague(League league) {
         return leagueRepository.save(league);
     }
 
     @Override
-    public void updateLeague(Long id, League league){
+    @Transactional
+    public void updateLeague(Long id, League league) {
         if (this.leagueRepository.findById(id).isPresent()) {
             League newLeague = this.leagueRepository.findById(id).get();
             newLeague.setName(league.getName());
