@@ -8,7 +8,7 @@ import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
 import { BehaviorSubject } from "rxjs";
 import { environment } from "../../../../../environments/environment";
 import { MatchStatus } from "../../../../enums/match-status.enum";
-import { League } from "../../../../models/league";
+import { League } from "../../../../models/league.model";
 import { Match } from "../../../../models/match";
 import { User } from "../../../../models/user";
 import { MatchService } from "../../../schedule/match.service";
