@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Subject } from "rxjs";
 import { delay } from "rxjs/operators";
-import { League } from "../../../models/league";
+import { League } from "../../../models/league.model";
 import { User } from "../../../models/user";
 import { UserService } from "../../administration/user-list/user.service";
 import { LeagueService } from "../../league/league.service";
@@ -52,7 +52,7 @@ export class UserMatchesComponent implements OnInit {
 
   loadLeagues() {
     this.isLoading = true;
-    this.leagueService.getLeagueList().pipe(delay(400)).subscribe(data => {
+    this.leagueService.getAllLeagues().pipe(delay(400)).subscribe(data => {
       this.leagueList = data;
       this.isLoading = false;
     });
